AI_Voice_Agent/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic[email]>=2.6
httpx>=0.26
openai>=1.40
google-api-python-client>=2.120
google-auth>=2.28
